refactor: HTML 관련 작업들 별도 파일로 분리
